removed all the static code added via hook #889
removed all the static code added via hook #889
