+++ templates/design-manager/design-1/frontpage.php
@@ -34,19 +34,17 @@ $template = new AMP_Post_Template( $post_id );?>
 		} ?>
 	<style amp-custom>
 	<?php $this->load_parts( array( 'style' ) ); ?>
-	<?php do_action( 'amp_post_template_css', $this ); ?>
+	<?php do_action( 'amp_post_template_css', $template ); ?>
 	</style>
 </head>
 <body class="single-post amp-wp-frontpage design_1_wrapper">
 	<?php $this->load_parts( array( 'header-bar' ) ); ?>
-	<?php do_action('ampforwp_frontpage_above_loop') ?>
-
-<?php ampforwp_frontpage_content($template, $post_id); ?>
+	<?php do_action('ampforwp_frontpage_above_loop',$template, $post_id) ?>
 	
 
-	<?php do_action('ampforwp_frontpage_below_loop') ?>
-	<?php do_action( 'amp_post_template_above_footer', $this ); ?>
+	<?php do_action('ampforwp_frontpage_below_loop',$template, $post_id) ?>
+	<?php do_action( 'amp_post_template_above_footer', $template ); ?>
 	<?php $this->load_parts( array( 'footer' ) ); ?>
-	<?php do_action( 'amp_post_template_footer', $this ); ?>
+	<?php do_action( 'amp_post_template_footer', $template ); ?>
 </body>
 </html>
