+++ templates/features.php
@@ -87,6 +87,8 @@
 	require 'custom-amp-content.php';
 	// Custom AMPFORWP Sanitizers
  	require 'custom-sanitizer.php';
+	// Custom Frontpage items
+ 	require 'frontpage-elements.php';
 //0.
 
 define('AMPFORWP_COMMENTS_PER_PAGE', $redux_builder_amp['ampforwp-number-of-comments'] );
@@ -2739,95 +2741,4 @@ function ampforwp_add_landing_page_elements() {
 function  ampforwp_remove_post_elements($elements) {
 	$elements =  array('empty-filter');
 	return $elements ;
-}
-
-
-
-function ampforwp_frontpage_content( $template, $post_id ){ 
-
-	global $redux_builder_amp;
-
-	$amp_custom_content_enable = get_post_meta($post_id, 'ampforwp_custom_content_editor_checkbox', true);
-
-	?>
-		<article class="amp-wp-article">
-
-		<?php if( $redux_builder_amp['ampforwp-title-on-front-page'] ) { ?>
-			<header class="amp-wp-article-header ampforwp-title">
-				<h1 class="amp-wp-title">
-						<?php echo get_the_title( $post_id ) ; ?>
-				</h1>
-			</header>
-		<?php } ?>
-
-		<?php do_action( 'ampforwp_after_header', $template );  ?>
-
-		<div class="amp-wp-content the_content">
-
-			<?php if (has_post_thumbnail( $post_id ) ):  ?>
-				<figure class="amp-wp-article-featured-image wp-caption"> <?php  
-					$image = wp_get_attachment_image_src( get_post_thumbnail_id( $post_id ), 'medium' ); 
-					$caption = get_the_post_thumbnail_caption( $post_id ); ?>
-					<amp-img src="<?php echo $image[0]; ?>" width="<?php echo $image[1]; ?>" height="<?php echo $image[2]; ?>" layout=responsive alt="<?php echo get_the_title( $post_id ) ?>" >  </amp-img>	
-					<?php if ( $caption ) : ?>
-						<p class="wp-caption-text">
-							<?php echo wp_kses_data( $caption ); ?>
-						</p>
-					<?php endif; ?>
-				</figure>
-			<?php endif; ?>
-		 <?php
-
-			do_action( 'ampforwp_before_post_content', $template );
-
-			// Normal Front Page Content
-			if ( ! $amp_custom_content_enable ) {
-				echo $template->get('post_amp_content');
-			} else {
-				// Custom/Alternative AMP content added through post meta
-				echo $template->get('ampforwp_amp_content');
-			}
-
-			do_action( 'ampforwp_after_post_content', $template );
-			?>
-
-		</div>
-
-		<?php ampforwp_frontpage_comments(); ?>
-
-		<div class="amp-wp-content post-pagination-meta">
-			<?php $template->load_parts( apply_filters( 'amp_post_template_meta_parts', array( 'meta-taxonomy' ) ) ); ?>
-		</div> <?php
-
-	
-		include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
-		if( $redux_builder_amp['enable-single-social-icons'] == true && !is_plugin_active( 'amp-cta/amp-cta.php' ) )  { ?>
-			<div class="sticky_social">
-				<?php if($redux_builder_amp['enable-single-facebook-share'] == true)  { ?>
-			    	<amp-social-share type="facebook"   width="50" height="28"></amp-social-share>
-			  	<?php } ?>
-			  	<?php if($redux_builder_amp['enable-single-twitter-share'] == true)  {
-						$data_param_data = $redux_builder_amp['enable-single-twitter-share-handle']; ?>
-						<amp-social-share type="twitter"
-							width="50"
-							height="28"
-							data-param-url=""
-							data-param-text="TITLE <?php echo wp_get_shortlink().' '.ampforwp_translation( $redux_builder_amp['amp-translator-via-text'], 'via' ).' '.$data_param_data ?>"
-						></amp-social-share>
-			  	<?php } ?>
-			  	<?php if($redux_builder_amp['enable-single-gplus-share'] == true)  { ?>
-			    	<amp-social-share type="gplus"      width="50" height="28"></amp-social-share>
-			  	<?php } ?>
-			  	<?php if($redux_builder_amp['enable-single-email-share'] == true)  { ?>
-			    	<amp-social-share type="email"      width="50" height="28"></amp-social-share>
-			  	<?php } ?>
-			  	<?php if($redux_builder_amp['enable-single-pinterest-share'] == true)  { ?>
-			    	<amp-social-share type="pinterest"  width="50" height="28"></amp-social-share>
-			  	<?php } ?>
-			  	<?php if($redux_builder_amp['enable-single-linkedin-share'] == true)  { ?>
-			    	<amp-social-share type="linkedin"   width="50" height="28"></amp-social-share>
-			  	<?php } ?>
-			</div>
-		<?php } ?>
-	</article> <?php 
 }
