+++ templates/frontpage-elements.php
@@ -1,0 +1,248 @@
+<?php 
+
+add_action('pre_amp_render_post', 'ampforwp_frontpage_file');
+function ampforwp_frontpage_file() {
+	global $redux_builder_amp;
+
+	if ( $redux_builder_amp['amp-design-selector'] == '1') {
+		add_action('ampforwp_frontpage_above_loop', 'ampforwp_design_1_frontpage_content', 10 , 2);
+	}
+	if ( $redux_builder_amp['amp-design-selector'] == '2') {
+		add_action('ampforwp_frontpage_above_loop', 'ampforwp_design_2_frontpage_content', 10 , 2);
+	}
+	if ( $redux_builder_amp['amp-design-selector'] == '3') {
+		add_action('ampforwp_frontpage_above_loop', 'ampforwp_design_3_frontpage_content', 10 , 2);
+	}
+}
+
+// Design #1
+function ampforwp_design_1_frontpage_content( $template, $post_id ){ 
+
+	global $redux_builder_amp;
+
+	$amp_custom_content_enable = get_post_meta($post_id, 'ampforwp_custom_content_editor_checkbox', true);
+
+	?>
+		<article class="amp-wp-article">
+
+		<?php if( $redux_builder_amp['ampforwp-title-on-front-page'] ) { ?>
+			<header class="amp-wp-article-header ampforwp-title">
+				<h1 class="amp-wp-title">
+						<?php echo get_the_title( $post_id ) ; ?>
+				</h1>
+			</header>
+		<?php } ?>
+
+		<?php do_action( 'ampforwp_after_header', $template );  ?>
+
+		<div class="amp-wp-content the_content">
+
+			<?php if (has_post_thumbnail( $post_id ) ):  ?>
+				<figure class="amp-wp-article-featured-image wp-caption"> <?php  
+					$image = wp_get_attachment_image_src( get_post_thumbnail_id( $post_id ), 'medium' ); 
+					$caption = get_the_post_thumbnail_caption( $post_id ); ?>
+					<amp-img src="<?php echo $image[0]; ?>" width="<?php echo $image[1]; ?>" height="<?php echo $image[2]; ?>" layout=responsive alt="<?php echo get_the_title( $post_id ) ?>" >  </amp-img>	
+					<?php if ( $caption ) : ?>
+						<p class="wp-caption-text">
+							<?php echo wp_kses_data( $caption ); ?>
+						</p>
+					<?php endif; ?>
+				</figure>
+			<?php endif; ?>
+		 <?php
+
+			do_action( 'ampforwp_before_post_content', $template );
+
+			// Normal Front Page Content
+			if ( ! $amp_custom_content_enable ) {
+				echo $template->get('post_amp_content');
+			} else {
+				// Custom/Alternative AMP content added through post meta
+				echo $template->get('ampforwp_amp_content');
+			}
+
+			do_action( 'ampforwp_after_post_content', $template );
+			?>
+
+		</div>
+
+		<?php ampforwp_frontpage_comments(); ?>
+
+		<div class="amp-wp-content post-pagination-meta">
+			<?php $template->load_parts( apply_filters( 'amp_post_template_meta_parts', array( 'meta-taxonomy' ) ) ); ?>
+		</div> <?php
+
+	
+		include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
+		if( $redux_builder_amp['enable-single-social-icons'] == true && !is_plugin_active( 'amp-cta/amp-cta.php' ) )  { ?>
+			<div class="sticky_social">
+				<?php if($redux_builder_amp['enable-single-facebook-share'] == true)  { ?>
+			    	<amp-social-share type="facebook"   width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-twitter-share'] == true)  {
+						$data_param_data = $redux_builder_amp['enable-single-twitter-share-handle']; ?>
+						<amp-social-share type="twitter"
+							width="50"
+							height="28"
+							data-param-url=""
+							data-param-text="TITLE <?php echo wp_get_shortlink().' '.ampforwp_translation( $redux_builder_amp['amp-translator-via-text'], 'via' ).' '.$data_param_data ?>"
+						></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-gplus-share'] == true)  { ?>
+			    	<amp-social-share type="gplus"      width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-email-share'] == true)  { ?>
+			    	<amp-social-share type="email"      width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-pinterest-share'] == true)  { ?>
+			    	<amp-social-share type="pinterest"  width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-linkedin-share'] == true)  { ?>
+			    	<amp-social-share type="linkedin"   width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			</div>
+		<?php } ?>
+	</article> <?php 
+}
+
+
+// Design #2
+function ampforwp_design_2_frontpage_content($template, $post_id){ ?>
+
+	<main>
+		<div class="amp-wp-content the_content"> 
+
+		<?php if (has_post_thumbnail( $post_id ) ):  ?>
+			<figure class="amp-wp-article-featured-image wp-caption"> <?php  
+				$image = wp_get_attachment_image_src( get_post_thumbnail_id( $post_id ), 'medium' ); 
+				$caption = get_the_post_thumbnail_caption( $post_id ); ?>
+				<amp-img src="<?php echo $image[0]; ?>" width="<?php echo $image[1]; ?>" height="<?php echo $image[2]; ?>" layout=responsive alt="<?php echo get_the_title( $post_id ) ?>" >  </amp-img>	
+				<?php if ( $caption ) : ?>
+					<p class="wp-caption-text">
+						<?php echo wp_kses_data( $caption ); ?>
+					</p>
+				<?php endif; ?>
+			</figure>
+		<?php endif; ?>	
+
+		<?php
+
+			// Normal Front Page Content
+			if ( ! $amp_custom_content_enable ) {
+				echo $template->get('post_amp_content');
+			} else {
+				// Custom/Alternative AMP content added through post meta
+				echo $template->get('ampforwp_amp_content');
+			}
+
+			do_action( 'ampforwp_after_post_content', $template ); ?>
+
+		</div>
+
+		<?php ampforwp_frontpage_comments(); ?>
+
+		<div class="amp-wp-content post-pagination-meta">
+			<?php $template->load_parts( apply_filters( 'amp_post_template_meta_parts', array( 'meta-taxonomy' ) ) ); ?>
+		</div> <?php
+
+
+		include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
+		if( $redux_builder_amp['enable-single-social-icons'] == true && !is_plugin_active( 'amp-cta/amp-cta.php' ) )  { ?>
+			<div class="sticky_social">
+				<?php if($redux_builder_amp['enable-single-facebook-share'] == true)  { ?>
+			    	<amp-social-share type="facebook"   width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-twitter-share'] == true)  { $data_param_data = $redux_builder_amp['enable-single-twitter-share-handle'];?>
+	      		<amp-social-share type="twitter"
+					width="50"
+					height="28"
+					data-param-url=""
+					data-param-text="TITLE <?php echo wp_get_shortlink().' '.ampforwp_translation( $redux_builder_amp['amp-translator-via-text'], 'via' ).' '.$data_param_data ?>"
+	      		></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-gplus-share'] == true)  { ?>
+			    	<amp-social-share type="gplus"      width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-email-share'] == true)  { ?>
+			    	<amp-social-share type="email"      width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-pinterest-share'] == true)  { ?>
+			    	<amp-social-share type="pinterest"  width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-linkedin-share'] == true)  { ?>
+			    	<amp-social-share type="linkedin"   width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			</div>
+		<?php } ?>
+	</main>
+
+	<?php 
+}
+
+// Design #3
+function ampforwp_design_3_frontpage_content($template, $post_id){ ?>
+	<main>
+		<div class="amp-wp-content the_content"> 
+
+			<?php if (has_post_thumbnail( $post_id ) ):  ?>
+				<figure class="amp-wp-article-featured-image wp-caption"> <?php  
+					$image = wp_get_attachment_image_src( get_post_thumbnail_id( $post_id ), 'medium' ); 
+					$caption = get_the_post_thumbnail_caption( $post_id ); ?>
+					<amp-img src="<?php echo $image[0]; ?>" width="<?php echo $image[1]; ?>" height="<?php echo $image[2]; ?>" layout=responsive alt="<?php echo get_the_title( $post_id ) ?>" >  </amp-img>	
+					<?php if ( $caption ) : ?>
+						<p class="wp-caption-text">
+							<?php echo wp_kses_data( $caption ); ?>
+						</p>
+					<?php endif; ?>
+				</figure>
+			<?php endif; ?>
+
+			<?php 
+			// Normal Front Page Content
+			if ( ! $amp_custom_content_enable ) {
+				echo $template->get('post_amp_content');
+			} else {
+				// Custom/Alternative AMP content added through post meta
+				echo $template->get('ampforwp_amp_content');
+			}
+
+			do_action( 'ampforwp_after_post_content', $template ); ?>
+
+		</div>
+
+		<?php ampforwp_frontpage_comments(); ?>
+
+		<div class="amp-wp-content post-pagination-meta">
+			<?php $template->load_parts( apply_filters( 'amp_post_template_meta_parts', array( 'meta-taxonomy' ) ) ); ?>
+		</div> <?php
+
+	  
+	  	include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
+	  	if( $redux_builder_amp['enable-single-social-icons'] == true && !is_plugin_active( 'amp-cta/amp-cta.php' ) )  { ?>
+			<div class="sticky_social">
+				<?php if($redux_builder_amp['enable-single-facebook-share'] == true)  { ?>
+			    	<amp-social-share type="facebook"   width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-twitter-share'] == true)  { $data_param_data = $redux_builder_amp['enable-single-twitter-share-handle'];?>
+			    	<amp-social-share type="twitter"    width="50" height="28"
+					data-param-url=""
+					data-param-text="TITLE <?php echo wp_get_shortlink().' '.ampforwp_translation( $redux_builder_amp['amp-translator-via-text'], 'via' ).' '.$data_param_data ?>">
+					</amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-gplus-share'] == true)  { ?>
+			    	<amp-social-share type="gplus"      width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-email-share'] == true)  { ?>
+			    	<amp-social-share type="email"      width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-pinterest-share'] == true)  { ?>
+			    	<amp-social-share type="pinterest"  width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			  	<?php if($redux_builder_amp['enable-single-linkedin-share'] == true)  { ?>
+			    	<amp-social-share type="linkedin"   width="50" height="28"></amp-social-share>
+			  	<?php } ?>
+			</div>
+		<?php } ?>
+	</main>
+	<?php 
+}
